+++ crates/ruma-common/CHANGELOG.md
@@ -24,6 +24,10 @@ Breaking changes:
     cargo feature allows to construct a `&'static EventId`.
   - `OwnedKeyId` was renamed to `KeyId`, and all its type aliases lost the `Owned` prefix too.
   - `OwnedAnyKeyName` was renamed to `AnyKeyName`.
+  - `OwnedMxcUri` was renamed to `MxcUri`. The `mxc_uri!` macro now returns an `MxcUri`, the
+    `owned_mxc_uri!` macro is behind the `unstable-identifier-owned-macros` cargo feature and
+    returns the same type. The `mxc_uri_ref!` macro behind the `unstable-identifier-ref-macros`
+    cargo feature still allows to construct a `&'static MxcUri`.
 
 
 Improvements: